Move page history views to wagtail.admin.views.pages.history
